+++ mainboard5/ssd1306.h
@@ -903,7 +903,7 @@ class IIC_DISPLAY {
     bool ssd1306_screen_exist = false;
     bool begin();
     bool screen_exist_test();
-    bool update(unsigned int x, unsigned int y, bool charge_or_not, bool charge_done_or_not, unsigned int battery_percent, unsigned int battery_voltage, bool screen_enabled, float backlight_duty_cycle, unsigned int fan_duty_cycle, unsigned int keyboard_status, unsigned int last_key);
+    bool update(unsigned int x, unsigned int y, bool charge_or_not, bool charge_done_or_not, unsigned int battery_percent, unsigned int battery_voltage, bool screen_enabled, uint8_t backlight_duty_cycle, uint8_t fan_duty_cycle, unsigned int keyboard_status, unsigned int last_key);
     bool flash();
 };
